pyproject.toml
[project]
name = "phishguard"
version = "0.1.0"
description = "Multi-modal phishing detector with late fusion."
authors = [{ name = "Solomon Smith" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "transformers>=4.40",
    "lightgbm>=4.3",
    "scikit-learn>=1.4",
    "pandas>=2.2",
    "numpy>=1.26",
    "tldextract>=5.1",
    "playwright>=1.43",
    "beautifulsoup4>=4.12",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "onnxruntime>=1.17",
    "pydantic>=2.6",
    "hydra-core>=1.3",
    "wandb>=0.16",
    "mlflow>=2.12",
    "evidently>=0.4",
    "pillow>=10.2",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "httpx>=0.27",
    "joblib>=1.3",
    "psycopg[binary]>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "ipykernel>=6.29",
    "jupyterlab>=4.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/phishguard/models/fusion.py" = ["N803"]
"src/phishguard/serving/api.py" = ["N806"]
"src/phishguard/training/train_url.py" = ["N806"]
"src/phishguard/training/train_fusion.py" = ["N806"]
"src/phishguard/training/train_screenshot.py" = ["N812"]
"src/phishguard/training/evaluate.py" = ["N806"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/phishguard --cov-report=term-missing"
markers = [
    "known_fail: expected failure for a documented, tracked issue -- treated as xfail",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
